rm_fix : remove extra newline when multi-layer shell receive SIGINT
