+++ src/parse/utils_parse_main.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils_main.c                                       :+:      :+:    :+:   */
+/*   utils_parse_main.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/14 12:32:30 by ghan              #+#    #+#             */
-/*   Updated: 2021/09/21 01:55:58 by ghan             ###   ########.fr       */
+/*   Updated: 2021/09/22 16:39:02 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
