+++ src/pipe/pipe_signal.c
@@ -3,26 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   pipe_signal.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yongjule <yongjule@42student.42seoul.      +#+  +:+       +#+        */
+/*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/18 11:22:44 by yongjule          #+#    #+#             */
-/*   Updated: 2021/09/22 12:25:06 by yongjule         ###   ########.fr       */
+/*   Updated: 2021/09/22 16:28:03 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "pipe.h"
 
-void	sigint_n_sigquit_handler(void (*sigfunction))
-{
-	signal(SIGINT, sigfunction);
-	signal(SIGQUIT, sigfunction);
-}
-
 void	signal_handle_wo_rl_prompt(int signal)
 {
-	if (signal == SIGINT)
+	if (signal == SIGINT || signal == SIGQUIT)
+	{
 		write(STDOUT_FILENO, "\n", 1);
-	return ;
+	}
 }
 
 void	reset_signal(int sig)
@@ -30,6 +25,15 @@ void	reset_signal(int sig)
 	if (sig == SIGINT || sig == SIGQUIT)
 	{
 		signal(sig, (void (*)(int))0);
+	}
+}
+
+void	multi_shell_erase_newline(int sig)
+{
+	if (sig == SIGINT)
+	{
+		rl_replace_line("", 0);
+		rl_redisplay();
 	}
 }
